Update the version on QNX custom resource table
When adding qnx_custom_rsc_table_vayu_ipu.h, the resource table version
was set to 2 which is not correct as the definition for the version is used
to differentiate the changes in the format of the resource table, not based
on its contents. Update the version of QNX custom resource table to 1.

Signed-off-by: Buddy Liong <a0270631@ti.com>