dce_register_engine shouldn't be called when Engine_open returns NULL
Prior to this patch, when Engine_open fails and return NULL Engine_Handle.
DCE will still caling dce_register_engine with NULL data which is not
necessary. The only issue here is that the c->refs count will be invalid.

This patch check on Engine_Handle to make sure it is not NULL before
calling dce_register_engine

Change-Id: I07680326ef991dbf32d9c7ceb2a1877bbd368f29
Signed-off-by: Buddy Liong <a0270631@ti.com>