dce_register_codec shouldn't be called when codec_create returns NULL
Prior to this patch, when codec_create fails and return NULL VIDDEC3_Handle.
DCE will still caling dce_register_codec with NULL data which is not
necessary. When dce_register_codec return < 0, it continue processing for
low latency mode, which is also not necessary.

This patch check on VIDDEC3_Handle to make sure it is not NULL before
calling dce_register_codec. It also moves the processing for low latency
mode to be executed only when dce_register_codec is successful.

Change-Id: I4b8fba03272c3d7c825a0bdbe813e74e04aaf555
Signed-off-by: Buddy Liong <a0270631@ti.com>