Run dce_init on Task thread
Previously dce_init() is configured as the BIOS UserStartupFunction or
started when BIOS_start() is called.
This is configured in \src\ti\framework\dce\package.xs

The assumption is that IpcMgr_rpmsgStartup() is also configured as the
BIOS UserStartupFunction or started when BIOS_start() is called.
This is configured in \platform\ti\dce\baseimage\dce_ipu.cfg

If IPC API IpcMgr_rpmsgStartup() is not configured as BIOS UserStartupFunction
but instead it was called from a Task thread, then dce_init has to be delayed
until RPMessage_init() which is called in IpcMgr_rpmsgStartup() is completed.
Otherwise rpmsg-dce and dce_callback thread in dce_init() will be failing
since the RPMSG is not configured yet.

This patch removes dce_init() from BIOS UserStartupFunction. It will now be
called from IPUMM_Main()

Change-Id: I6d2cb8f4cf55d72c25eb867ad48ed7e2710efb4b
Signed-off-by: Buddy Liong <a0270631@ti.com>