Invalidating parameter buffers for codec_get_version
Previously version_buf pointer is set to buffer pointer (data.buf)
of (IVIDDEC3_Status *) status.
The pointer dyn_params and status are not cache invalidated first which
can cause the status->data.buf to be accessed without cache invalidation.
Due to this corruption, it can cause exception on IPUMM.

The solution is to cache invalidate the dyn_params and status pointer
first to get the valid pointer of dyn_params and status.
Then set the version_buf with status->data.buf.
Perform cache invalidation on the version_buf to get the valid pointer.

Change-Id: I3258cca0ee191c488fc6457391215b8bfcde3b85
Signed-off-by: Subash Lakkimsetti <x0091084@ti.com>