Add profiler to measure based on IVA-HD processing time
When running with QNX test app, we are not able to get a good data
on IVA fps because the Profiler calculation is including the time at MPU
side. With QNX test app, there are more time spend on MPU side due to
file reading to an input buffer.

Previously the calculation is only based on the Total time which is the time
on both MPU and IVA.
This commit adds additional profiler to be based only on the IVA processing time.
The total IVA processing time is calculated based on the accumulation of time before
calling codec and after codec process is completed.

Change-Id: Ie754b156faa8d8ba508981d5b74bdb70518517cb
Signed-off-by: Buddy Liong <a0270631@ti.com>